ENH: Reduim el cost de memòria de cada imatge a 2 bytes ( signed short int ). Fins ara feiem servir 4 bytes ( signed int ). Segons DICOM no s'ha trobat cap modalitat que necessiti més de 16 bits.

+++ starviewer/src/core/volume.h
@@ -46,7 +46,7 @@ class Volume : public QObject
 Q_OBJECT
 public:
     /// Tipus d'imatge intern per defecte d'itk
-    typedef signed int ItkPixelType;
+    typedef signed short int ItkPixelType;
     static const unsigned int VDimension = 3;
 
     typedef itk::Image<ItkPixelType, VDimension >    ItkImageType;
